requirements.txt
numpy
panda
tensorflow=1.13.1
ampligraph
plotly
matplotlib